refactor: extract shared patch machinery from fix.go into patch.go
Prepares for the compile command to reuse uniqueMatch, applyEdits,
and contentChecksum without duplicating the byte-level edit logic.
